+++ relay-profiling/src/cocoa.rs
@@ -4,7 +4,8 @@ use serde::{de, Deserialize, Serialize};
 
 use relay_general::protocol::{Addr, DebugId, EventId, NativeImagePath};
 
-use crate::utils::deserialize_number_from_string;
+use crate::transaction_metadata::TransactionMetadata;
+use crate::utils::{deserialize_number_from_string, is_zero};
 use crate::ProfileError;
 
 fn strip_pointer_authentication_code<'de, D>(deserializer: D) -> Result<Addr, D::Error>
@@ -16,13 +17,13 @@ where
     Ok(Addr(addr.0 & 0x0000000FFFFFFFFF))
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 struct Frame {
     #[serde(deserialize_with = "strip_pointer_authentication_code")]
     instruction_addr: Addr,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 struct Sample {
     frames: Vec<Frame>,
 
@@ -36,7 +37,7 @@ struct Sample {
     thread_id: u64,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 struct ThreadMetadata {
     #[serde(default)]
     is_main_thread: bool,
@@ -48,12 +49,12 @@ struct ThreadMetadata {
     priority: Option<u32>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 struct QueueMetadata {
     label: String,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 struct SampledProfile {
     samples: Vec<Sample>,
 
@@ -64,14 +65,14 @@ struct SampledProfile {
     queue_metadata: HashMap<String, QueueMetadata>,
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 #[serde(rename_all = "lowercase")]
 enum ImageType {
     MachO,
     Symbolic,
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 pub struct NativeDebugImage {
     #[serde(alias = "name")]
     code_file: NativeImagePath,
@@ -89,12 +90,12 @@ pub struct NativeDebugImage {
     image_type: ImageType,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 struct DebugMeta {
     images: Vec<NativeDebugImage>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 struct CocoaProfile {
     debug_meta: DebugMeta,
     device_is_emulator: bool,
@@ -105,23 +106,44 @@ struct CocoaProfile {
     device_os_name: String,
     device_os_version: String,
 
-    #[serde(deserialize_with = "deserialize_number_from_string")]
-    duration_ns: u64,
-
     #[serde(default, skip_serializing_if = "String::is_empty")]
     environment: String,
 
     platform: String,
     profile_id: EventId,
     sampled_profile: SampledProfile,
-    trace_id: EventId,
-    transaction_id: EventId,
-    transaction_name: String,
     version_code: String,
     version_name: String,
+
+    #[serde(default, skip_serializing_if = "EventId::is_nil")]
+    transaction_id: EventId,
+    #[serde(default, skip_serializing_if = "String::is_empty")]
+    transaction_name: String,
+    #[serde(default, skip_serializing_if = "EventId::is_nil")]
+    trace_id: EventId,
+    #[serde(
+        default,
+        deserialize_with = "deserialize_number_from_string",
+        skip_serializing_if = "is_zero"
+    )]
+    duration_ns: u64,
+
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    transactions: Vec<TransactionMetadata>,
 }
 
 impl CocoaProfile {
+    fn set_transaction(&mut self, transaction: &TransactionMetadata) {
+        self.transaction_name = transaction.name.clone();
+        self.transaction_id = transaction.id;
+        self.trace_id = transaction.trace_id;
+        self.duration_ns = transaction.relative_end_ns - transaction.relative_start_ns;
+    }
+
+    fn has_transaction_metadata(&self) -> bool {
+        !self.transaction_name.is_empty() && self.duration_ns > 0
+    }
+
     /// Removes a sample when it's the only sample on its thread
     fn remove_single_samples_per_thread(&mut self) {
         let mut sample_count_by_thread_id: HashMap<u64, u32> = HashMap::new();
@@ -141,7 +163,50 @@ impl CocoaProfile {
     }
 }
 
-pub fn parse_cocoa_profile(payload: &[u8]) -> Result<Vec<u8>, ProfileError> {
+pub fn expand_cocoa_profile(payload: &[u8]) -> Result<Vec<Vec<u8>>, ProfileError> {
+    let profile = parse_cocoa_profile(payload)?;
+    let mut items: Vec<Vec<u8>> = Vec::new();
+
+    if profile.transactions.is_empty() && profile.has_transaction_metadata() {
+        match serde_json::to_vec(&profile) {
+            Ok(payload) => items.push(payload),
+            Err(_) => {
+                return Err(ProfileError::CannotSerializePayload);
+            }
+        };
+
+        return Ok(items);
+    }
+
+    for transaction in &profile.transactions {
+        let mut new_profile = profile.clone();
+
+        new_profile.set_transaction(transaction);
+        new_profile.transactions.clear();
+
+        new_profile.sampled_profile.samples.retain_mut(|sample| {
+            if transaction.relative_start_ns <= sample.relative_timestamp_ns
+                && sample.relative_timestamp_ns <= transaction.relative_end_ns
+            {
+                sample.relative_timestamp_ns -= transaction.relative_start_ns;
+                true
+            } else {
+                false
+            }
+        });
+
+        match serde_json::to_vec(&new_profile) {
+            Ok(payload) => items.push(payload),
+            Err(_) => {
+                return Err(ProfileError::CannotSerializePayload);
+            }
+        };
+    }
+
+    Ok(items)
+}
+
+fn parse_cocoa_profile(payload: &[u8]) -> Result<CocoaProfile, ProfileError> {
     let mut profile: CocoaProfile =
         serde_json::from_slice(payload).map_err(ProfileError::InvalidJson)?;
 
@@ -151,10 +216,17 @@ pub fn parse_cocoa_profile(payload: &[u8]) -> Result<Vec<u8>, ProfileError> {
         return Err(ProfileError::NotEnoughSamples);
     }
 
-    match serde_json::to_vec(&profile) {
-        Ok(payload) => Ok(payload),
-        Err(_) => Err(ProfileError::CannotSerializePayload),
+    if profile.transactions.is_empty() && !profile.has_transaction_metadata() {
+        return Err(ProfileError::NoTransactionAssociated);
     }
+
+    for transaction in profile.transactions.iter() {
+        if !transaction.valid() {
+            return Err(ProfileError::InvalidTransactionMetadata);
+        }
+    }
+
+    Ok(profile)
 }
 
 #[cfg(test)]
@@ -166,9 +238,10 @@ mod tests {
 
     #[test]
     fn test_roundtrip_cocoa() {
-        let payload = include_bytes!("../tests/fixtures/profiles/cocoa.json");
-        let data = parse_cocoa_profile(payload);
-        assert!(data.is_ok());
+        let payload = include_bytes!("../tests/fixtures/profiles/cocoa/valid.json");
+        let profile = parse_cocoa_profile(payload);
+        assert!(profile.is_ok());
+        let data = serde_json::to_vec(&profile.unwrap());
         assert!(parse_cocoa_profile(&data.unwrap()[..]).is_ok());
     }
 
@@ -190,6 +263,20 @@ mod tests {
                 image_vmaddr: Annotated::new(Addr(4294967296)),
                 other: Map::new(),
             }))), annotated);
+    }
+
+    #[test]
+    fn test_parse_multiple_transactions() {
+        let payload = include_bytes!("../tests/fixtures/profiles/cocoa/multiple_transactions.json");
+        let data = parse_cocoa_profile(payload);
+        assert!(data.is_ok());
+    }
+
+    #[test]
+    fn test_no_transaction() {
+        let payload = include_bytes!("../tests/fixtures/profiles/cocoa/no_transaction.json");
+        let data = parse_cocoa_profile(payload);
+        assert!(data.is_err());
     }
 
     fn generate_profile() -> CocoaProfile {
@@ -214,6 +301,7 @@ mod tests {
             trace_id: EventId::new(),
             transaction_id: EventId::new(),
             transaction_name: "test".to_string(),
+            transactions: Vec::new(),
             version_code: "9999".to_string(),
             version_name: "1.0".to_string(),
         }
@@ -286,7 +374,29 @@ mod tests {
         ]);
 
         let payload = serde_json::to_vec(&profile).unwrap();
-
         assert!(parse_cocoa_profile(&payload[..]).is_err());
+    }
+
+    #[test]
+    fn test_expand_multiple_transactions() {
+        let payload = include_bytes!("../tests/fixtures/profiles/cocoa/multiple_transactions.json");
+        let data = expand_cocoa_profile(payload);
+        assert!(data.is_ok());
+        assert_eq!(data.as_ref().unwrap().len(), 2);
+
+        let profile = match parse_cocoa_profile(&data.as_ref().unwrap()[0][..]) {
+            Err(err) => panic!("cannot parse profile: {:?}", err),
+            Ok(profile) => profile,
+        };
+        assert_eq!(
+            profile.transaction_id,
+            "30976f2ddbe04ac9b6bffe6e35d4710c".parse().unwrap()
+        );
+        assert_eq!(profile.duration_ns, 8);
+        assert_eq!(profile.sampled_profile.samples.len(), 2);
+
+        for sample in &profile.sampled_profile.samples {
+            assert!(sample.relative_timestamp_ns < profile.duration_ns);
+        }
     }
 }